INITIAL COMMIT
This is the initial commit for an adapted version of the neural network
that was used in the MIT STAG project (http://stag.csail.mit.edu/).
The adapted version allows for testing of the same architecture with
reduced complexity.
Additionally, the training/test loss and accuracy are logged to
visualize their evolution over epochs.